An http client implementation from scratch
I basically implemented an http client completely from scratch. It can
only be used for the specific purpose of downloading a single resource
and dosn't support most of the features of http such as: content
compression, chunked encoding, keep-alive etc.

This implementation however is implemented using non-blocking
SocketChannels witch means we will be able to run several downloads in
parallel from a single thread.

I still need to make content available to the caller (and not written to
the file system).